Move view tests into test_views
